requirements.txt
flask
requests
together
python-dotenv
gunicorn
mysql-connector-python
fuzzywuzzy 
python-Levenshtein
pymysql
